[CHORE] JWT 예외 메세지 구체화 #156
