+++ src/main/java/com/example/namoldak/util/jwt/JwtAuthFilter.java
@@ -1,8 +1,5 @@
 package com.example.namoldak.util.jwt;
 
-import com.example.namoldak.util.GlobalResponse.GlobalResponseDto;
-import com.example.namoldak.util.GlobalResponse.code.StatusCode;
-import com.fasterxml.jackson.databind.ObjectMapper;
 import io.jsonwebtoken.Claims;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
@@ -28,20 +25,11 @@ public class JwtAuthFilter extends OncePerRequestFilter {
         String token = jwtUtil.resolveToken(request);
 
         if(token != null) {
-            if(!jwtUtil.validateToken(token)){
-                response.setStatus(StatusCode.INVALID_TOKEN.getHttpStatus().value());
-                response.setContentType("application/json; charset=UTF-8");
-                try {
-                    String json = new ObjectMapper().writeValueAsString(new GlobalResponseDto(StatusCode.INVALID_TOKEN));
-                    response.getWriter().write(json);
-                } catch (Exception e) {
-                    log.error(e.getMessage());
-                }
-                return;
+            if(jwtUtil.validateToken(token)){
+                // 3. 토큰이 유효하다면 토큰에서 정보를 가져와 Authentication 에 세팅
+                Claims info = jwtUtil.getUserInfoFromToken(token);
+                setAuthentication(info.getSubject());
             }
-            // 3. 토큰이 유효하다면 토큰에서 정보를 가져와 Authentication 에 세팅
-            Claims info = jwtUtil.getUserInfoFromToken(token);
-            setAuthentication(info.getSubject());
         }
         // 4. 다음 필터로 넘어간다
         filterChain.doFilter(request, response);
@@ -52,15 +40,5 @@ public class JwtAuthFilter extends OncePerRequestFilter {
         Authentication authentication = jwtUtil.createAuthentication(loginId);
         context.setAuthentication(authentication);
         SecurityContextHolder.setContext(context);
-    }
-    public void jwtExceptionHandler(HttpServletResponse response, String message, int statusCode) {
-        response.setStatus(statusCode);
-        response.setContentType("application/json");
-        try {
-            String json = new ObjectMapper().writeValueAsString(new GlobalResponseDto<>(StatusCode.INVALID_TOKEN));
-            response.getWriter().write(json);
-        } catch (Exception e) {
-            log.error(e.getMessage());
-        }
     }
 }
